+++ src/test/java/lotto/domain/LottoPurchaseServiceTest.java
@@ -1,4 +1,4 @@
-package lotto.service;
+package lotto.domain;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
@@ -12,7 +12,7 @@ class LottoPurchaseServiceTest {
     @DisplayName("로또 개수 구하기(1000원 이상)")
     void 로또_개수_구하기_1000원_이상(){
 
-        LottoPurchaseService lotto = new LottoPurchaseService();
+        Lottos lotto = new Lottos();
 
         assertThat(lotto.calculateNumberOfLotto(14000)).isEqualTo(14);
 
@@ -21,7 +21,7 @@ class LottoPurchaseServiceTest {
     @Test
     @DisplayName("1000원 미만 입력 시 예외 발생")
     void 로또_구매_불가_1000원_미만(){
-        LottoPurchaseService lottoPurchaseService = new LottoPurchaseService();
+        Lottos lottoPurchaseService = new Lottos();
         Assertions.assertThatThrownBy(
                 () -> lottoPurchaseService.calculateNumberOfLotto(999)
         ).isInstanceOf(IllegalArgumentException.class);
